Implement q_sort()
Use merge sort to achieve O(nlongn) time.
The idea is take the list as a single link list and use merge
sort.
So we first set the end of the list to zero,and call merge_sort(),
merge_sort() for devide  the list to half(use slow fast pointer),
and conquer the list use merge() to merge left and right.
By the way i use count variable to determind whether which half is
the head of the answer.